requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
matrix-client==0.0.5
patreon==0.5.0
patroniser==0.3
PyYAML==3.12
requests==2.18.4
urllib3==1.22

# pip install https://github.com/lampholder/matrix-python-sdk/archive/master.zip
